docker/registry-mgr/requirements.txt
python-dxf >= 9.0.1 
absl-py
pyhcl
